requirements.txt
pytest==6.1.2
requests==2.22.0
selenium==3.141.0
pytest-html==3.1.1
allure-pytest==2.9.43
